[Away] Give some more flair to listening and gaming
This gives a little bit more flair to the listening and gaming options providing details about the users activity such as:
Game details if the game provides it or a thumbnail.

Listening shows the current placement in the song just like the audio cog as well as the album art now.

You can also now set an image as the response if embeds are enabled and the image is a valid embed image.

This also fixes a bug when the bot restarts removing old settings and a bug with replacing mentions in listening status.