[RSS] Make RSS work with cog disabling API (#160)
* [RSS] Make RSS work with cog disabling API

* Update feed qualifier to work with disabling

Instead of the catch-all dealing with feeds having no time, just let the catch-all handle everything. If the cog has been disabled for a long period of time and there are no matches via the saved feed information, the feed will only post 1 feed post instead of all of them (10, 20, 25 posts depending on the feed usually). If the cog is reinstated and there is only a partial match (say, 18 out of 20 feeds before it finds a qualifying match)... it will still post all 18 posts.

Co-authored-by: aikaterna <20862007+aikaterna@users.noreply.github.com>